Add s2_sample_mm & Move samples to a sub folder
