--- sample_mm/sample_mm.vdf
@@ -1,5 +1,0 @@
-"Metamod Plugin"
-{
-	"alias"		"sample"
-	"file"		"addons/sample_mm"
-}
