+++ samples/s1_sample_mm/AMBuildScript
