+++ samples/s1_sample_mm/sample_mm.vdf
@@ -1,0 +1,5 @@
+"Metamod Plugin"
+{
+	"alias"		"sample"
+	"file"		"addons/sample_mm"
+}
