+++ samples/s1_stub_mm/README.md
@@ -1,0 +1,4 @@
+For more information on compiling and reading the plugin's source code, see:
+
+	http://wiki.alliedmods.net/Category:Metamod:Source_Development
+
