+++ samples/s1_stub_mm/stub_mm.vdf
@@ -1,0 +1,5 @@
+"Metamod Plugin"
+{
+	"alias"		"stub"
+	"file"		"addons/stub_mm"
+}
