--- stub_mm/README.txt
@@ -1,4 +1,0 @@
-For more information on compiling and reading the plugin's source code, see:
-
-	http://wiki.alliedmods.net/Category:Metamod:Source_Development
-
