--- stub_mm/stub_mm.vdf
@@ -1,5 +1,0 @@
-"Metamod Plugin"
-{
-	"alias"		"stub"
-	"file"		"addons/stub_mm"
-}
